Beefed up error checking (massively)
